fix and refactor tests
